[fix] #14 주문 생성 로직 수정
이전에는 클라이언트로부터 주문 상품에 대한 내용을 받았었음
장바구니에서 사용자와 매장 번호로 데이터를 가져와 주문 후 삭제하도록 수정

+++ backend/src/main/java/com/spicy/backend/order/dto/response/OrderItemResponse.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record OrderItemResponse(
         String productName,
 
-        Integer quantity,
+        Long quantity,
 
         BigDecimal unitPrice,
